+++ errors/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To be Released
 
+* feat(UnwrapError) `UnwrapError` now unwraps errors which implement an `Unwrap()` method.
+
 ## v3.0.0
 
 * fix(errors): `Build` returns `error` not `*ValidationErrors` (BREAKING CHANGE)